* Implemented -keepinpackindex * Changed extension of humanreadable indexes from ttt to txt * Removed obsoloted autoclicker data * More detailed "Already unpacked" message in log * Fixed a bug which caused FAR to be hidden * -showdrpnames is saved in cfg * Implemented -extractdir:<dir> * Fixed -log_dir * Fixed manager_restorepos * Implemented -extractonly * showdrpnames no longer displays full path

+++ trunk/source/theme.c
@@ -471,7 +471,7 @@ void theme_enum(HWND hwnd,WCHAR *path)
     if(!(FindFileData.dwFileAttributes&FILE_ATTRIBUTE_DIRECTORY))
     {
         wsprintf(buf,L"%s\\%s\\%s",data_dir,path,FindFileData.cFileName);
-        D(THEME_NAME)=L"";
+        D(THEME_NAME)=(int)L"";
         vault_loadfromfile(buf,theme,&theme_strs,&themedata);
         SendMessage(hwnd,CB_ADDSTRING,0,(int)D(THEME_NAME));
         wcscpy(themelist[i],buf);
